requirements.txt
absl-py==2.3.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.18.4
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.2
asgiref==3.11.0
astunparse==1.6.3
attrs==25.1.0
Automat==24.8.1
av==14.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.3.0
blinker==1.8.2
Brotli==1.1.0
build==1.4.0
cachetools==6.2.2
certifi==2025.7.14
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.1
chromadb==1.5.1
click==8.3.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
ConfigArgParse==1.7.1
constantly==23.10.4
contourpy==1.3.2
cors==1.0.1
cryptography==46.0.5
ctranslate2==4.6.0
cycler==0.12.1
Cython==3.1.2
dataclasses-json==0.6.7
datasets==4.2.0
decorator==5.2.1
deepface==0.0.96
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
dm-tree==0.1.9
dnspython==2.7.0
docstring_parser==0.17.0
durationpy==0.10
easydict==1.13
ecdsa==0.19.1
einops==0.8.2
email-validator==2.3.0
etils==1.13.0
facenet-pytorch==2.6.0
Farama-Notifications==0.0.4
fastapi==0.110.0
fastembed==0.7.4
faster-whisper==1.1.1
filelock==3.18.0
fire==0.7.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flatbuffers==25.2.10
flutter==0.1
fonttools==4.58.4
frozenlist==1.8.0
fsspec==2025.5.1
future==1.0.0
gast==0.7.0
gaze-tracking==0.0.1
gdown==5.2.1
geopandas==1.1.2
gevent==24.11.1
geventhttpclient==2.3.4
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
googleapis-common-protos==1.72.0
googletrans==4.0.2
greenlet==3.1.1
groq==1.0.0
grpcio==1.76.0
gTTS==2.5.4
gunicorn==21.2.0
gymnasium==1.2.0
h11==0.16.0
h2==3.2.0
h5py==3.15.1
hf-xet==1.1.3
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.1
humanfriendly==10.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.37.0
imbalanced-learn==0.14.1
immutabledict==4.3.1
importlib_metadata==8.7.1
importlib_resources==6.5.2
incremental==24.7.2
iniconfig==2.0.0
insightface==0.7.3
itsdangerous==2.1.2
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.2
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==2.0.0
kagglesdk==0.1.16
keras==3.13.0
kiwisolver==1.4.8
kubernetes==35.0.0
langchain==0.2.5
langchain-classic==1.0.1
langchain-community==0.2.5
langchain-core==0.2.8
langchain-huggingface==1.2.0
langchain-openai==1.1.7
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.1
langsmith==0.1.81
lazy_loader==0.4
libclang==18.1.1
lightecc==0.0.4
lightgbm==4.6.0
lightphe==0.0.20
limits==5.4.0
llvmlite==0.45.1
locust==2.37.10
locust-cloud==1.23.2
loguru==0.7.3
lxml==5.4.0
lz4==4.4.5
Mako==1.3.10
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mmh3==5.2.0
mne==1.10.2
mongoengine==0.27.0
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.12.0
networkx==3.5
nltk==3.8.1
numba==0.62.1
numpy==1.26.4
oauthlib==3.3.1
onnx==1.18.0
onnxruntime==1.22.0
openai==2.15.0
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opencv-python-headless==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.18.0
ordered-set==4.1.0
orjson==3.11.5
osmnx==2.1.0
overrides==7.7.0
packaging==24.2
pandas==2.1.4
passlib==1.7.4
patsy==1.0.2
pillow==10.4.0
pinecone==7.3.0
pinecone-client==5.0.1
pinecone-plugin-assistant==1.8.0
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
plotly==6.5.0
pluggy==1.5.0
pooch==1.8.2
posthog==5.4.0
prettytable==3.16.0
promise==2.3
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.0.0
psycopg2-binary==2.9.10
py_rust_stemmers==0.1.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyjson==1.4.1
PyJWT==2.10.1
pymongo==4.6.1
PyMySQL==1.1.1
pynndescent==0.5.13
pyogrio==0.12.1
pyparsing==3.2.3
pypdf==6.6.0
pypickle==1.1.0
PyPika==0.51.1
pyproj==3.7.2
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.2
python-jose==3.3.0
python-multipart==0.0.21
python-pptx==1.0.2
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.2.3
sentencepiece==0.2.0
setuptools==75.8.2
shapely==2.1.2
shellingham==1.5.4
simple-parsing==0.1.8
simple-websocket==1.1.0
simsimd==6.5.0
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sqlparse==0.5.5
stable_baselines3==2.7.0
starlette==0.36.3
statsmodels==0.14.5
streamlit==1.51.0
stringzilla==3.12.5
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.3
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.12.0
tldextract==5.1.3
tokenizers==0.22.1
toml==0.10.2
torch==2.2.2
torchaudio==2.7.1
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
Twisted==24.11.0
typer==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
uritemplate==4.2.0
urllib3==2.3.0
uuid_utils==0.14.0
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
Werkzeug==2.3.7
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xgboost==3.0.5
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.event==5.0
zope.interface==7.2
zstandard==0.25.0
